Moved GOPanelView to the gui folder
