+++ src/grandorgue/GODocument.cpp
@@ -15,6 +15,7 @@
 #include "dialogs/midi-event/GOMidiEventDialog.h"
 #include "document-base/GOView.h"
 #include "gui/GOGUIPanel.h"
+#include "gui/GOGUIPanelView.h"
 #include "midi/GOMidiEvent.h"
 #include "sound/GOSound.h"
 #include "threading/GOMutexLocker.h"
@@ -23,7 +24,6 @@
 #include "GOFrame.h"
 #include "GOOrgan.h"
 #include "GOOrganController.h"
-#include "GOPanelView.h"
 #include "GOResizable.h"
 #include "go_ids.h"
